+++ token.go
@@ -1,59 +1,83 @@
+/*
+	Спасибо Артему за идею с TokenAgent
+*/
+
 package finam
 
 import (
 	"context"
-	"fmt"
+	"sync"
 	"time"
 
-	pb "github.com/Ruvad39/go-finam-grpc/proto/grpc/tradeapi/v1/auth"
-	"google.golang.org/grpc/metadata"
+	auth_service "github.com/Ruvad39/go-finam-grpc/proto/grpc/tradeapi/v1/auth"
 )
 
-const authKey = "Authorization"             //
-const jwtTokenTtl = 12 * time.Minute        // Время жизни токена JWT в минутах (15 минут)
-const jwtRefreshInterval = 10 * time.Minute // Интервал обновления токена (в минутах)
+const authKey = "Authorization"            //
+const jwtRefreshInterval = 9 * time.Minute // Интервал обновления токена (в минутах)
+
+// TokenAgent
+// соответствует интерфейсу PerRPCCredentials
+type TokenAgent struct {
+	jwt string // JWT токен для дальнейшей авторизации
+	mu  sync.RWMutex
+}
+
+func NewTokenAgent() *TokenAgent {
+	return &TokenAgent{}
+}
+
+func (a *TokenAgent) setJwt(jwt string) {
+	a.mu.Lock()
+	defer a.mu.Unlock()
+	a.jwt = jwt
+}
+
+func (a *TokenAgent) GetJwt() string {
+	a.mu.RLock()
+	defer a.mu.RUnlock()
+	return a.jwt
+}
+
+// GetRequestMetadata для соответствия интерфейсу PerRPCCredentials
+func (a *TokenAgent) GetRequestMetadata(ctx context.Context, uri ...string) (map[string]string, error) {
+	meta := make(map[string]string)
+	meta[authKey] = a.GetJwt()
+
+	return meta, nil
+}
+
+// GetRequestMetadata для соответствия интерфейсу PerRPCCredentials
+func (a *TokenAgent) RequireTransportSecurity() bool {
+	return true
+}
 
 // GetJWT Получение JWT токена из API токена
 //
 // идет  вызов AuthService.Auth
 func (c *Client) GetJWT(ctx context.Context) (string, error) {
 	if c.token == "" {
-		c.accessToken = ""
-		return c.accessToken, nil
+		c.TokenAgent.setJwt("")
+		return "", nil
 	}
-	req := &pb.AuthRequest{Secret: c.token}
+	req := &auth_service.AuthRequest{Secret: c.token}
 	log.Debug("GetJWT start AuthService.Auth")
 	t := time.Now()
 	res, err := c.AuthService.Auth(ctx, req)
 	if err != nil {
-		return c.accessToken, err
+		return "", err
 	}
 	log.Debug("GetJWT end AuthService.Auth", "duration", time.Since(t))
 	return res.Token, nil
 }
 
-// UpdateJWT
-// если jwt токен пустой или вышло его время
-// получим JWT
-// и запишем его в параметры клиента. Проставим время получения
-func (c *Client) UpdateJWT(ctx context.Context) error {
-	if c.token == "" {
-		c.accessToken = ""
-		return fmt.Errorf("UpdateJWT: token пустой")
+// JwtRefresh обновим токен
+func (c *Client) JwtRefresh(ctx context.Context) error {
+	token, err := c.GetJWT(ctx)
+	if err != nil {
+		log.Error("failed to refresh JW", "err", err.Error())
+		return err
 	}
-	// если токен пустой или вышло его время
-	if c.accessToken == "" || c.ttlJWT.Before(time.Now()) {
-		// получим новый токен
-		log.Debug("UpdateJWT. токен пустой или вышло его время = получим новый токен")
-		token, err := c.GetJWT(ctx)
-		if err != nil {
-			return err
-		}
-		// запишем время окончания токена
-		c.ttlJWT = time.Now().Add(jwtTokenTtl)
-		c.accessToken = token
-	}
-	log.Debug("UpdateJWT. токен живой")
+	c.TokenAgent.setJwt(token)
 	return nil
 }
 
@@ -61,24 +85,18 @@ func (c *Client) UpdateJWT(ctx context.Context) error {
 func (c *Client) runJwtRefresher(ctx context.Context) {
 	log.Debug("run JwtRefresher")
 	if c.token == "" {
-		c.accessToken = ""
+		c.TokenAgent.setJwt("")
 		log.Warn("JWT refresher token пустой. Выход из метода")
 		return
 	}
-	ticker := time.NewTicker(jwtRefreshInterval)
+	ticker := time.NewTicker(c.opts.jwtRefreshInterval)
 	defer ticker.Stop()
 
 	for {
 		select {
 		case <-ticker.C:
 			log.Debug("start JwtRefresh")
-			token, err := c.GetJWT(ctx)
-			if err != nil {
-				log.Error("failed to refresh JW", "err", err.Error())
-			}
-			// запишем время окончания токена
-			c.ttlJWT = time.Now().Add(jwtTokenTtl)
-			c.accessToken = token
+			_ = c.JwtRefresh(ctx)
 		case <-ctx.Done():
 			log.Debug("JWT refresher stopped")
 			return
@@ -89,22 +107,6 @@ func (c *Client) runJwtRefresher(ctx context.Context) {
 // GetTokenDetails Получение информации о токене сессии
 //
 // идет вызов AuthService.TokenDetails
-func (c *Client) GetTokenDetails(ctx context.Context) (*pb.TokenDetailsResponse, error) {
-	return c.AuthService.TokenDetails(ctx, &pb.TokenDetailsRequest{Token: c.accessToken})
-}
-
-// WithAuthToken Создаем новый контекст с заголовком Authorization
-// пишем в него jwt токен
-func (c *Client) WithAuthToken(ctx context.Context) (context.Context, error) {
-	// проверим наличие токена
-	err := c.UpdateJWT(ctx)
-	if err != nil {
-		return ctx, err
-	}
-	// добавим заголовок
-	md := metadata.New(map[string]string{
-		authKey: c.accessToken,
-	})
-	// и добавляем его в ctx
-	return metadata.NewOutgoingContext(ctx, md), nil
+func (c *Client) GetTokenDetails(ctx context.Context) (*auth_service.TokenDetailsResponse, error) {
+	return c.AuthService.TokenDetails(ctx, &auth_service.TokenDetailsRequest{Token: c.TokenAgent.GetJwt()})
 }